# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/faits": {
      "post": {
        "description": "Un lien créé ici est lisible depuis ses deux extrémités : une seule arête est stockée.",
        "operationId": "FaitsController_creer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreationFaitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaitDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Ajout d’un fait à une entité existante",
        "tags": [
          "faits"
        ]
      }
    },
    "/faits/{id}": {
      "patch": {
        "description": "La cible d’un lien ne se corrige pas : un lien mal posé s’infirme et se refait.",
        "operationId": "FaitsController_modifier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificationFaitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaitDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Correction d’un fait",
        "tags": [
          "faits"
        ]
      }
    },
    "/faits/{id}/infirmer": {
      "post": {
        "description": "Un fait contredit sort du graphe actif et reste consultable dans l’onglet Historique. Il n’est jamais supprimé — ce que le service a cru un moment fait partie de ce qu’il a su.",
        "operationId": "FaitsController_infirmer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfirmationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaitDto"
                }
              }
            }
          },
          "409": {
            "description": "Déjà infirmé"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Infirmation",
        "tags": [
          "faits"
        ]
      }
    },
    "/entites/{id}/fichiers": {
      "get": {
        "description": "Les images d’une fiche inaccessible sont introuvables — 404, jamais 403.",
        "operationId": "FichiersController_lister",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FichierDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Images d’une entité",
        "tags": [
          "fichiers"
        ]
      },
      "post": {
        "description": "Le type est vérifié **sur le contenu** et non sur l’extension, la taille est plafonnée, et les métadonnées — EXIF, GPS, commentaires — sont retirées sans réencoder l’image.",
        "operationId": "FichiersController_deposer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "fichier": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FichierDto"
                }
              }
            }
          },
          "400": {
            "description": "Format refusé ou fichier vide"
          },
          "413": {
            "description": "Au-delà du plafond de taille"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Dépôt d’une image",
        "tags": [
          "fichiers"
        ]
      }
    },
    "/fichiers/{id}": {
      "get": {
        "description": "Aucun dossier n’est servi en statique : les droits sont vérifiés avant que l’octet ne parte.",
        "operationId": "FichiersController_telecharger",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "L’image"
          },
          "404": {
            "description": "Inconnu, ou fiche inaccessible"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Téléchargement authentifié",
        "tags": [
          "fichiers"
        ]
      }
    },
    "/graphe": {
      "get": {
        "description": "Voisinage d’une entité, élagué pour cet agent **avant** traversée. Chaque nœud rapporte le nombre de voisins qu’il reste à déplier.",
        "operationId": "GrapheController_voisinage",
        "parameters": [
          {
            "name": "depuis",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "profondeur",
            "required": false,
            "in": "query",
            "description": "Nombre de sauts, 1 à 4",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "fiabilite",
            "required": false,
            "in": "query",
            "description": "Niveau minimal des arêtes retenues, 1 à 4",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "dossier",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoisinageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Exploration par expansion",
        "tags": [
          "graphe"
        ]
      }
    },
    "/graphe/complet": {
      "get": {
        "description": "Tout ce que l’agent peut voir, à toute profondeur, élagué **avant** constitution. Les fiches isolées y figurent : une entité sans lien fait partie de ce que le service sait.",
        "operationId": "GrapheController_vueEntiere",
        "parameters": [
          {
            "name": "fiabilite",
            "required": false,
            "in": "query",
            "description": "Niveau minimal des arêtes retenues, 1 à 4",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "dossier",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoisinageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Vue entière",
        "tags": [
          "graphe"
        ]
      }
    },
    "/graphe/chemin": {
      "get": {
        "description": "Deux résultats : le trajet le plus court et le plus solide. Un chemin passant par un lien masqué n’existe pas pour l’agent concerné — la plateforme ne dit jamais qu’un chemin existe mais reste inaccessible.",
        "operationId": "GrapheController_chemin",
        "parameters": [
          {
            "name": "de",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "vers",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "fiabilite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheminsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Recherche de chemin",
        "tags": [
          "graphe"
        ]
      }
    },
    "/graphe/positions": {
      "post": {
        "description": "Par dossier lorsqu’un dossier est indiqué, globalement sinon. Le repositionnement affecte tous les agents : c’est une disposition partagée, d’où la permission.",
        "operationId": "GrapheController_enregistrerPositions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispositionDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Disposition mémorisée",
        "tags": [
          "graphe"
        ]
      }
    },
...
  },
  "info": {
    "title": "Centrale N&I",
    "description": "API du service Narcotics & Investigations de la SAHP. Le référentiel — types d'entités, champs, types de liens, onglets — est dynamique et n'apparaît pas dans ce contrat : il se récupère par GET /referentiel.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jeton": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreationFaitDto": {
        "type": "object",
        "properties": {
          "sujetId": {
            "type": "string",
            "format": "uuid"
          },
          "nature": {
            "enum": [
              "champ",
              "lien"
            ],
            "type": "string"
          },
          "definitionChampId": {
            "type": "string",
            "format": "uuid",
            "description": "Si nature = champ"
          },
          "valeur": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              }
            ],
            "description": "Si nature = champ"
          },
          "typeLienId": {
            "type": "string",
            "format": "uuid",
            "description": "Si nature = lien"
          },
          "cibleId": {
            "type": "string",
            "format": "uuid",
            "description": "Si nature = lien"
          },
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "example": "Rapport d'intervention n°2291"
          },
          "fiabilite": {
            "type": "number",
            "minimum": 1,
            "maximum": 4
          },
          "dateConstatation": {
            "type": "string",
            "format": "date",
            "example": "2026-08-07"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "dossierId": {
            "type": "string",
            "format": "uuid",
            "description": "Dossier de saisie — le fait en hérite la visibilité"
          }
        },
        "required": [
          "sujetId",
          "nature",
          "source",
          "fiabilite",
          "dateConstatation"
        ]
      },
      "FaitDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sujetId": {
            "type": "string",
            "format": "uuid"
          },
          "nature": {
            "enum": [
              "champ",
              "lien"
            ],
            "type": "string"
          },
          "definitionChampId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "valeur": {
            "nullable": true,
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "array",
                "items": {}
              }
            ]
          },
          "typeLienId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "cibleId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "source": {
            "type": "string"
          },
          "fiabilite": {
            "type": "number"
          },
          "dateConstatation": {
            "type": "string",
            "format": "date"
          },
          "etat": {
            "enum": [
              "actif",
              "infirme",
              "archive"
            ],
            "type": "string"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "visibiliteEffective": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string",
            "description": "Calculée en base : la plus restrictive des quatre gardiens"
          },
          "dossierId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "creeLe": {
            "type": "string",
            "format": "date-time"
          },
          "modifieLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "sujetId",
          "nature",
          "definitionChampId",
          "valeur",
          "typeLienId",
          "cibleId",
          "source",
          "fiabilite",
          "dateConstatation",
          "etat",
          "visibilite",
          "visibiliteEffective",
          "dossierId",
          "creeLe",
          "modifieLe"
        ]
      },
      "ModificationFaitDto": {
        "type": "object",
        "properties": {
          "valeur": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              }
            ],
            "description": "Champ seulement — la cible d’un lien ne se corrige pas"
          },
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "fiabilite": {
            "type": "number",
            "minimum": 1,
            "maximum": 4
          },
          "dateConstatation": {
            "type": "string",
            "format": "date"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          }
        }
      },
      "InfirmationDto": {
        "type": "object",
        "properties": {
          "motif": {
            "type": "string",
            "minLength": 3,
            "maxLength": 512,
            "description": "Ce qui contredit le fait"
          }
        },
        "required": [
          "motif"
        ]
      },
      "FichierDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "entiteId": {
            "type": "string",
            "format": "uuid"
          },
          "nomOrigine": {
            "type": "string",
            "description": "Nom du fichier tel que l’agent l’a déposé"
          },
          "mime": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          },
          "taille": {
            "type": "number",
            "description": "Taille après retrait des métadonnées, en octets"
          },
          "deposeLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "entiteId",
          "nomOrigine",
          "mime",
          "taille",
          "deposeLe"
        ]
      },
      "NoeudGrapheDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string"
          },
          "typeCode": {
            "type": "string"
          },
          "typeEntiteId": {
            "type": "string",
            "format": "uuid"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "voisinsNonAffiches": {
            "type": "number",
            "description": "Voisins atteignables mais non affichés — le badge d’expansion. Ne compte jamais ce qui est masqué à cet agent."
          },
          "recurrence": {
            "type": "boolean",
            "description": "Relie plusieurs entités appartenant à des dossiers différents"
          },
          "x": {
            "type": "number",
            "nullable": true
          },
          "y": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "id",
          "libelle",
          "typeCode",
          "typeEntiteId",
          "visibilite",
          "voisinsNonAffiches",
          "recurrence",
          "x",
          "y"
        ]
      },
      "AreteGrapheDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identifiant du fait"
          },
          "sujetId": {
            "type": "string",
            "format": "uuid"
          },
          "cibleId": {
            "type": "string",
            "format": "uuid"
          },
          "typeLienId": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string"
          },
          "fiabilite": {
            "type": "number",
            "minimum": 1,
            "maximum": 4
          }
        },
        "required": [
          "id",
          "sujetId",
          "cibleId",
          "typeLienId",
          "libelle",
          "fiabilite"
        ]
      },
      "VoisinageDto": {
        "type": "object",
        "properties": {
          "noeuds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoeudGrapheDto"
            }
          },
          "aretes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AreteGrapheDto"
            }
          }
        },
        "required": [
          "noeuds",
          "aretes"
        ]
      },
      "CheminDto": {
        "type": "object",
        "properties": {
          "noeuds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoeudGrapheDto"
            }
          },
          "aretes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AreteGrapheDto"
            }
          },
          "longueur": {
            "type": "number",
            "description": "Nombre de sauts"
          },
          "maillonLeFaible": {
            "type": "number",
            "minimum": 1,
            "maximum": 4,
            "description": "La fiabilité du chemin est celle de son maillon le plus faible"
          }
        },
        "required": [
          "noeuds",
          "aretes",
          "longueur",
          "maillonLeFaible"
        ]
      },
      "CheminsDto": {
        "type": "object",
        "properties": {
          "plusCourt": {
            "nullable": true,
            "description": "Le trajet le plus court. Nul si aucun chemin n’existe.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CheminDto"
              }
            ]
          },
          "plusSolide": {
            "nullable": true,
            "description": "Le trajet dont le maillon le plus faible est le plus élevé. Nul lorsqu’il coïncide avec le plus court — un seul est alors affiché.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CheminDto"
              }
            ]
          }
        },
        "required": [
          "plusCourt",
          "plusSolide"
        ]
      },
      "PositionDto": {
        "type": "object",
        "properties": {
          "entiteId": {
            "type": "string",
            "format": "uuid"
          },
          "x": {
            "type": "number"
          },
          "y": {
            "type": "number"
          }
        },
        "required": [
          "entiteId",
          "x",
          "y"
        ]
      },
      "DispositionDto": {
        "type": "object",
        "properties": {
          "dossierId": {
            "type": "string",
            "format": "uuid",
            "description": "Disposition propre à ce dossier. Absent : disposition globale."
          },
          "positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionDto"
            }
          }
        },
        "required": [
          "positions"
        ]
      },
...
    }
  }
}